+++ apps/supplement-stack/CHANGELOG.md
@@ -1,5 +1,18 @@
 # supplement-stack
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/graphql@0.0.27
+  - @zemble/core@0.1.2
+  - zemble-plugin-auth-otp@0.0.25
+  - zemble-plugin-bull@0.0.27
+  - zemble-plugin-todo@0.0.23
+  - zemble-plugin-apple-app-site-association@0.0.18
+  - @zemble/bun@0.0.13
+
 ## 0.0.24
 
 ### Patch Changes