+++ packages/auth-api-token/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zemble-plugin-auth-api-token
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/graphql@0.0.27
+  - @zemble/core@0.1.2
+  - zemble-plugin-auth@0.0.25
+
 ## 0.0.24
 
 ### Patch Changes