+++ packages/email-sendgrid/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zemble-plugin-email-sendgrid
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/graphql@0.0.27
+  - @zemble/core@0.1.2
+  - zemble-plugin-kv@0.0.25
+
 ## 0.0.24
 
 ### Patch Changes