+++ packages/kv/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zemble-plugin-kv
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/graphql@0.0.27
+  - @zemble/core@0.1.2
+
 ## 0.0.24
 
 ### Patch Changes